backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.0.5
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
plotly==5.15.0
opencv-python==4.8.0.76
Pillow==10.0.0
gunicorn==21.2.0
python-dotenv==1.0.0
pytest==7.4.0
requests==2.31.0